Fix error handling for xml to json and, mulitpart parser
Both the handlers were setting the response code as `400` and sending
the response straight from the middleware. This causes issues in
debugging as these errors are never logged anywhere, and the only place
where we can find a trace of these is in nginx access logs etc., where the error
messages are lost and we can only see the response codes.

Since ExpressJS follows error first callback patterns, we can pass the
caught erro/exception to `next` as the first argument, and let the end-user
decide how to handle the error.

This follows the pattern followed by the original `body-parser`
wherein `next` is called with an error if parsing fails for some reason.

Ideally middleware should only be transforming/analysing requests and
not serving them, hence moving the response handling out of the
body-parser.